+++ src/trans.c
@@ -97,7 +97,7 @@ static NTSTATUS trans_commit(device_extension* Vcb, trans_ref* trans) {
 
             if (fr->created) {
                 ExAcquireResourceExclusiveLite(fr->parent->fcb->Header.Resource, true);
-                fr->parent->fcb->inode_item.st_size += fr->dc->utf8.Length * 2;
+                fr->parent->fcb->inode_item.st_size += (uint64_t)fr->dc->utf8.Length * 2;
                 fr->parent->fcb->inode_item.sequence++;
                 fr->parent->fcb->inode_item_changed = true;
                 ExReleaseResourceLite(fr->parent->fcb->Header.Resource);
@@ -107,7 +107,7 @@ static NTSTATUS trans_commit(device_extension* Vcb, trans_ref* trans) {
                 mark_fcb_dirty(fr->parent->fcb);
             } else if (fr->oldutf8.Buffer) { // renamed
                 ExAcquireResourceExclusiveLite(fr->parent->fcb->Header.Resource, true);
-                fr->parent->fcb->inode_item.st_size = fr->parent->fcb->inode_item.st_size + (2 * fr->dc->utf8.Length) - (2 * fr->oldutf8.Length);
+                fr->parent->fcb->inode_item.st_size = fr->parent->fcb->inode_item.st_size + (2 * (uint64_t)fr->dc->utf8.Length) - (2 * (uint64_t)fr->oldutf8.Length);
                 fr->parent->fcb->inode_item.sequence++;
                 fr->parent->fcb->inode_item_changed = true;
                 ExReleaseResourceLite(fr->parent->fcb->Header.Resource);
@@ -116,7 +116,7 @@ static NTSTATUS trans_commit(device_extension* Vcb, trans_ref* trans) {
             }
         } else if (fr->deleted && fr->oldutf8.Buffer) {
             ExAcquireResourceExclusiveLite(fr->parent->fcb->Header.Resource, true);
-            fr->parent->fcb->inode_item.st_size -= fr->oldutf8.Length * 2;
+            fr->parent->fcb->inode_item.st_size -= (uint64_t)fr->oldutf8.Length * 2;
             fr->parent->fcb->inode_item.sequence++;
             fr->parent->fcb->inode_item_changed = true;
             ExReleaseResourceLite(fr->parent->fcb->Header.Resource);
